error fix--initiate default tab
This tab shall be showing the dashboard as has always been as most users love/ prefer to look at their shinny stuff all in one screen. This is to be fixed tomorrow.
Also, the existing tabs shall remain and statuses are to be fixed the following day. on the issue of if or not a user can have one or more form records assigned to them, I got tips from Mubarak, shall work upon them the following day